add postmapping annotation for adding headers in request add api responses annotations and get movie from feign client

+++ src/main/java/dclvs/moviehubapi/clients/KinopoiskUnofficialAPIClient.java
@@ -9,6 +9,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface KinopoiskUnofficialAPIClient {
 
     @GetMapping("${kinopoisk.unofficial.api.path}/{id}")
-    MovieResponse getMovieById(@PathVariable("id") Long id);
+    MovieResponse getMovieById(@PathVariable("id") Integer id);
 
 }
